+++ tests/testthat/test-install.R
@@ -1,0 +1,4 @@
+test_that("install java and check", {
+  install_jdk()
+  testthat::expect_true(java_available())
+})