+++ NEWS.md
@@ -1,4 +1,10 @@
+# multilinguer 0.0.7
+
+# multilinguer 0.0.6
+
+* Add install_conda() with reticulate package.
+
 # multilinguer 0.0.4
 
-* [hotfix] Fixed download corretto link.
-* Added a `NEWS.md` file to track changes to the package.
+* [hotfix] Fix download corretto link.
+* Add a `NEWS.md` file to track changes to the package.